+++ app/assets/src/components/ui/controls/search_box_list.scss
@@ -6,7 +6,7 @@
   display: inline-block;
   font-size: 13px;
   min-width: 200px;
-  padding: 20px;
+  padding: 10px;
   text-align: left;
 
   .searchBox {
