+++ app/assets/src/components/views/report/SampleDetailsSidebar/SampleDetailsSidebar.jsx
@@ -15,12 +15,8 @@ import {
 import MetadataTab from "./MetadataTab";
 import PipelineTab from "./PipelineTab";
 import NotesTab from "./NotesTab";
-import {
-  processMetadata,
-  processMetadataTypes,
-  processPipelineInfo,
-  processAdditionalInfo
-} from "../../../utils/metadata";
+import { processMetadata, processMetadataTypes } from "../../../utils/metadata";
+import { processPipelineInfo, processAdditionalInfo } from "./info_utils";
 import cs from "./sample_details_sidebar.scss";
 
 const TABS = ["Metadata", "Pipeline", "Notes"];
